requirements.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
Deprecated==1.2.14
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
smmap==5.0.1
typing_extensions==4.9.0
urllib3==2.1.0
wrapt==1.16.0
